refactor(SelectDropdown): react-select pkg bump (#2300)
* hacky thing

* real yarn

* reorg and start typing

* correctly typed react-select, need to fix label accessibility

* actually typed, still need to fix label

* labels working, cool

* updated packages

* alright alright alright

* test passing

* rename type

* searchable

* steady

* yarn test

* yarn test

* icons fixed

* style tweak

* colormode

* remove unused imports

* => get option bg

* merge
