+++ packages/gamut/package.json
@@ -30,7 +30,6 @@
     "@reach/auto-id": "^0.16.0",
     "@reach/tabs": "^0.16.4",
     "@types/marked": "^1.1.0",
-    "@types/react-select": "^4.0.13",
     "classnames": "^2.2.5",
     "framer-motion": "^4.1.17",
     "html-to-react": "^1.4.2",
@@ -42,7 +41,7 @@
     "react-focus-on": "^3.5.1",
     "react-hook-form": "^7.21.2",
     "react-player": "^2.3.1",
-    "react-select": "^4.3.0",
+    "react-select": "^5.2.2",
     "react-truncate-markup": "^5.1.0",
     "react-use": "^15.3.8",
     "sanitize-markdown": "^2.6.6"
