+++ packages/gamut/src/Form/SelectDropdown/index.tsx
@@ -1,0 +1,2 @@
+export * from './SelectDropdown';
+export * from './types';
