+++ packages/gamut/src/Form/styles/index.ts
@@ -1,4 +1,4 @@
 export * from './shared-system-props';
 export * from './Checkbox-styles';
 export * from './Radio-styles';
-export * from './SelectDropdown-styles';
+export * from '../SelectDropdown/styles';
